+++ addon/services/moment.js
@@ -1,9 +1,9 @@
 import Ember from 'ember';
 import moment from 'moment';
 
-const { computed, get, set, Logger:logger } = Ember;
+const { computed, get, getProperties, set, Logger:logger } = Ember;
 
-export default Ember.Service.extend({
+export default Ember.Service.extend(Ember.Evented, {
   _timeZone: null,
 
   locale: null,
@@ -16,7 +16,7 @@ export default Ember.Service.extend({
 
     set(propertyKey, timeZone) {
       if (!moment.tz) {
-        logger.warn('[ember-moment] attempted to set timezone, but moment-timezone unavailable.');
+        logger.warn('[ember-moment] attempted to set timezone, but moment-timezone is not setup.');
         return;
       }
 
@@ -26,12 +26,23 @@ export default Ember.Service.extend({
     }
   }),
 
+  setLocale(locale) {
+    this.changeLocale(locale);
+  },
+
   changeLocale(locale) {
     set(this, 'locale', locale);
+    moment.locale(locale);
+    this.trigger('localeChanged', locale);
+  },
+
+  setTimeZone(timeZone) {
+    this.changeTimeZone(timeZone);
   },
 
   changeTimeZone(timeZone) {
     set(this, 'timeZone', timeZone);
+    this.trigger('timeZoneChanged', timeZone);
   },
 
   isMoment(obj) {
@@ -40,8 +51,7 @@ export default Ember.Service.extend({
 
   moment() {
     let momentObj = moment(...arguments);
-    const locale = get(this, 'locale');
-    const timeZone = get(this, 'timeZone');
+    let { locale, timeZone } = getProperties(this, 'locale', 'timeZone');
 
     if (locale && momentObj.locale) {
       momentObj = momentObj.locale(locale);