+++ tests/integration/moment-test.js
@@ -5,8 +5,8 @@ import { moduleForComponent, test } from 'ember-qunit';
 moduleForComponent('moment', {
   integration: true,
   beforeEach() {
-    this.momentService = this.container.lookup('service:moment');
-    this.momentService.changeLocale('en');
+    this.service = this.container.lookup('service:moment');
+    this.service.changeLocale('en');
   }
 });
 
@@ -14,8 +14,8 @@ test('moment-from and moment integration', function(assert) {
   assert.expect(1);
 
   this.set('tomorrow', moment().add(1, 'day'));
-
   this.render(hbs`{{moment-from (moment) tomorrow}}`);
+
   assert.equal(this.$().text(), 'a day ago');
 });
 
@@ -23,8 +23,8 @@ test('moment-from and moment integration', function(assert) {
   assert.expect(1);
 
   this.set('tomorrow', moment().add(1, 'day'));
-
   this.render(hbs`{{moment-to (moment) tomorrow}}`);
+
   assert.equal(this.$().text(), 'in a day');
 });
 
@@ -41,7 +41,7 @@ test('moment and monent-format helper integration #2', function(assert) {
   assert.equal(this.$().text(), 'May 3, 2010');
 });
 
-test('moment can use the global locale', function(assert) {
+test('moment can use the service locale', function(assert) {
   assert.expect(1);
 
   this.setProperties({
@@ -50,7 +50,47 @@ test('moment can use the global locale', function(assert) {
     date: '5/3/10'
   });
 
-  this.momentService.changeLocale('fr');
+  this.service.changeLocale('fr');
   this.render(hbs`{{moment-format (moment date inputFormat) outputFormat}}`);
+
+  assert.equal(this.$().text(), 'mai 3, 2010');
+});
+
+test('changing moment service locale changes global locale', function(assert) {
+  const done = assert.async();
+  assert.expect(1);
+
+  this.service.on('localeChanged', function() {
+    assert.equal(moment.locale(), 'es');
+    done();
+  });
+
+  this.service.setLocale('es');
+});
+
+test('changing timeZone triggers event', function(assert) {
+  const done = assert.async();
+  assert.expect(1);
+
+  this.service.on('timeZoneChanged', function() {
+    assert.ok(true);
+    done();
+  });
+
+  this.service.setTimeZone('PST');
+});
+
+test('moment can use the service locale (setLocale)', function(assert) {
+  assert.expect(1);
+
+  this.setProperties({
+    inputFormat: 'M/D/YY',
+    outputFormat: 'MMMM D, YYYY',
+    date: '5/3/10'
+  });
+
+  this.service.setLocale('fr');
+  this.render(hbs`{{moment-format (moment date inputFormat) outputFormat}}`);
+
   assert.equal(this.$().text(), 'mai 3, 2010');
 });